refactor(utils): move methods to global utils folder
