+++ src/utils/processFullName.ts
@@ -1,0 +1,4 @@
+export const processFullName = (fullName: string) => {
+  const [firstName, lastName] = fullName.split(' ');
+  return { firstName, lastName };
+};
